% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_series_num.R
\name{check_series_num}
\alias{check_series_num}
\title{Check Series Number}
\usage{
check_series_num(num, current, PT)
}
\arguments{
\item{num}{A character vector containing the series details.}

\item{PT}{A data frame that contains the periodic table
information.}
}
\description{
A function that tries to parse and format the
 series that is being investigated into an integer value.
}
