% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_details.R
\name{element_details}
\alias{element_details}
\title{Element Details}
\usage{
element_details(currentElement, PT)
}
\description{
A function to check whether an element is
 in the periodic table or not.
}
