% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx_Functions.R
\name{Gtex_Sample_Finder}
\alias{Gtex_Sample_Finder}
\title{Gtex_Sample_Finder Function}
\usage{
Gtex_Sample_Finder(Tissue)
}
\arguments{
\item{Tissue}{The Tissue of Interest as per GTEx Portal}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
Gtex_Sample_Finder(Tissue="Whole_Blood");
}
