% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx_Functions.R
\name{GTEx_Summary}
\alias{GTEx_Summary}
\title{GTEx Summary Function}
\usage{
GTEx_Summary(S_G_T_Matrix)
}
\arguments{
\item{S_G_T_Matrix}{A dataframe in the format of SNP GENE TISSUE to be provides}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
S_G_T_Matrix <- data.frame(SNP = c("chr3_49771630" , "chr3_49772408" , "chr3_49773528" , "chr3_49773614" , "chr3_49773845") , Gene = c("ENSG00000182179.12" , "ENSG00000182179.12" ,"ENSG00000182179.12" , "ENSG00000182179.12" ,"ENSG00000182179.12") , Tissue = c("Whole_Blood","Lung","Liver","Pancreas","Ovary"))
GTEx_Summary(S_G_T_Matrix)
}
