% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nist.R
\name{nist}
\alias{nist}
\title{NIST}
\usage{
nist(INPUT)
}
\arguments{
\item{INPUT}{A named list contaning the input flags from
running the script, INPUT can contain the following
  \itemize{
    \item{\code{element} : An element to query for e.g. fe ii}
    \item{\code{conversion} : One of ev / ryd / icm}
    \item{\code{overwrite} : TRUE or FALSE}
  }}
}
\description{
A function that handles all the components of
 querying for data based on the input provided. Files and
 data are written from here.
}
