% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nist_query.R
\name{nist_query}
\alias{nist_query}
\title{Get data}
\usage{
nist_query(series, INPUT)
}
\arguments{
\item{series}{A character string containing an ionic series
which is of the form \code{{Element} {Series}}, e.g. Co ii.}
}
\description{
A function that sends a post query to NIST to
 retrieve the HTTP response containing the level information
 in the form of a table, nested in an XML nodeset.
}
