% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_checker.R
\name{parse_series}
\alias{parse_series}
\title{Parse Series}
\usage{
parse_series(series)
}
\arguments{
\item{series}{A character string containing an ionic series.}
}
\description{
A function that sends a post query to NIST to
 retrieve the HTTP response containing the level information
 in the form of a table, nested in an XML nodeset.
}
\details{
This function checks to see if a series is valid or
 not, and can take multiple formats and parses it into something
 valid for sending the query. It probably works either way, but
 I can at least let this fail early if it doesn't look the way it
 should. Element and series number formats can consist of
   \itemize{
     \item{\code{Helium / helium / He / he}}
     \item{\code{2 / 1 / ii / i}}
   }
}
