% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx_Functions.R
\name{snp_1kgp}
\alias{snp_1kgp}
\title{snp_1kgp Function}
\usage{
snp_1kgp(chr, start, end)
}
\arguments{
\item{chr}{The Tissue of Interest as per GTEx Portal}

\item{start}{The Tissue of Interest as per GTEx Portal}

\item{end}{The Tissue of Interest as per GTEx Portal}
}
\value{
set of snps
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
snp_1kgp(chr = 1 , start = 13500 , end = 14000)
}
