% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx_Functions.R
\name{snp_gtex}
\alias{snp_gtex}
\title{Gtex_Sample_Finder Function}
\usage{
snp_gtex(chr, start, end)
}
\arguments{
\item{Tissue}{The Tissue of Interest as per GTEx Portal}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
snp_gtex(chr = 1, start = 13526, end = 14000)
}
