% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haploreg_Functions.R
\name{snpToAF}
\alias{snpToAF}
\title{snpToAF Function}
\usage{
snpToAF(rsids)
}
\arguments{
\item{rsids}{A vector of rsids}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
Retrieve allele frequencies for a set of SNPs from the HAPLOREG database.
}
\examples{
snpToAF(rsids = "rs10")
snpToAF(rsids = c("rs10","rs3","rs4","rs10047249"))
}
