% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haploreg_Functions.R
\name{snpToCHROMHMM_25STATE}
\alias{snpToCHROMHMM_25STATE}
\title{snpToCHROMHMM_25STATE Function}
\usage{
snpToCHROMHMM_25STATE(rsid)
}
\arguments{
\item{rsid}{A single variant with rsid}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
Retrieve the CHROMHMM 25-state annotations for a set of SNPs from the HAPLOREG database.
}
\examples{
snpToCHROMHMM_25STATE(rsid <- "rs10");
}
