% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haploreg_Functions.R
\name{snpToMotif}
\alias{snpToMotif}
\title{snpToMotif Function}
\usage{
snpToMotif(rsid)
}
\arguments{
\item{rsid}{genotypic variant as rsid}
}
\value{
Summary Statistics of the SNP,Gene and Tissue
}
\description{
Retrieve MOTIF information for a set of SNPs from the HAPLOREG database.
}
\examples{
snpToMotif(rsid = "rs10")
}
