% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx_Functions.R
\name{snp_gtex_signif}
\alias{snp_gtex_signif}
\title{snp_gtex_signif Function}
\usage{
snp_gtex_signif(gene = "ENSG00000225630", tissue = "Liver")
}
\arguments{
\item{gene}{The Tissue of Interest as per GTEx Portal}

\item{tissue}{The Tissue of Interest as per GTEx Portal}
}
\value{
set of snps
}
\description{
A function in R that retrieves a summary of SNP-gene-tissue associations from the Genotype-Tissue Expression GTEx dataset.
}
\examples{
snp_gtex_signif(gene = "ENSG00000225630" , tissue= "Liver")
}
